requirements.txt
fastapi
mangum
uvicorn
numpy
python-dotenv
google-generativeai
openai
tqdm
beautifulsoup4
requests
bs4